chore: move approval hook helpers into multi op trade confirm hooks directory
